fix offer item structure.

+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Money.java
@@ -1,7 +1,11 @@
 package pl.com.bottega.ecommerce.sales.domain.offer;
 
+import java.math.BigDecimal;
+
 public class Money {
-    public double value;
+    public BigDecimal value;
 
     public String currency;
+
+
 }
